refactor: main에 작성됐던 GameController의 로직과 InputView의 로직을 각각 분리
